refactor(db): extract indexer write path to db/src/sync.rs
