+++ db/src/lib.rs
@@ -18,6 +18,7 @@ pub mod pool;
 pub mod queries;
 pub mod scanner;
 pub mod settings;
+pub mod sync;
 mod taxonomy;
 pub mod thumbs;
 pub mod worker;
@@ -37,4 +38,5 @@ pub use metadata_overrides::*;
 pub use pool::*;
 pub use queries::*;
 pub use settings::*;
+pub use sync::*;
 pub use thumbs::{thumb_path_for, thumbs_dir, ThumbError, ThumbSize};
